% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{AGHN}
\alias{AGHN}
\title{Fraga Alves, Gomes, de Haan, Neves (2009) Mixed Moment Estimator}
\usage{
AGHN(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the mixed moment tail index estimator from Fraga Alves, Gomes, de Haan, and Neves (2009), as described on page 3.
}
\details{
This estimator refines the tail index using mixed moments and an auxiliary bias adjustment:
- The first moment \( M_1 \) is computed using the \code{MM} function with \( \alpha = 1 \).
- A bias correction term \( L_1 \) is calculated as:
  \[
  L_1 = \text{mean}\left( 1 - \frac{X_{(n-k)}}{X_{(n-i+1)}} \right), \, i = 1, \ldots, k.
  \]
- The auxiliary parameter \( \phi \) is then calculated as:
  \[
  \phi = \frac{M_1 - L_1}{L_1^2}.
  \]
- The tail index is computed as:
  \[
  \text{Tail Index} = \frac{1 + 2 \cdot \min(\phi - 1, 0)}{\phi - 1}.
  \]
- The adjustment ensures non-positive bias corrections to improve accuracy.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
AGHN(x)

}
