% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{Alves}
\alias{Alves}
\title{Alves (1995) Tail Index Estimator}
\usage{
Alves(X, k, k0)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}

\item{k0}{An optional integer specifying the number of order statistics used for the auxiliary calculation. Defaults to \( k^{3/4} \).}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the tail index estimator from Alves (1995), suitable for both positive and negative tail indexes.
}
\details{
This estimator allows for the calculation of both positive and negative tail indexes by analyzing ratios of differences between order statistics:
- The auxiliary parameter \( k_0 \) controls the subset of top-order statistics for the ratio calculation and defaults to \( k^{3/4} \).
- The tail index is calculated as:
  \[
  \text{Tail Index} = \left[ \frac{1}{k_0} \sum_{i=1}^{k_0} \log\left(\frac{X_{(n-k_0+i)} - X_{(n-k)}}{X_{(n-k_0)} - X_{(n-k)}}\right) \right]^{-1}.
  \]

This method provides a flexible approach for estimating tail indexes, including those with negative values.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
Alves(x)

}
