% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{Alves1995}
\alias{Alves1995}
\title{Alves (1995) Tail Index Estimator}
\usage{
Alves1995(X, k, c)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}

\item{c}{An optional numeric constant to control the scaling parameter. Defaults to \code{5}.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the tail index estimator from Alves (1995), suitable for positive tail indexes.
}
\details{
This estimator computes the tail index using the log ratio of order statistics:
- The constant \( c \) scales the order statistics used in the calculation. If \( c \cdot k > n \), it is adjusted to \( c = 2 \).
- The tail index is calculated as:
  \[
  \text{Tail Index} = \frac{\log(c)}{\log\left(\frac{X_{(n-k+1)}}{X_{(n-c \cdot k + 1)}}\right)}.
  \]
This method is designed for positive tail indexes and adjusts the scaling dynamically if the scaling factor exceeds the sample size.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
Alves1995(x)

}
