% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{am}
\alias{am}
\title{\insertCite{aban2001;textual}{TIE} Shifted Hill's Estimator}
\usage{
am(X, k, eps = 1e-07)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}

\item{eps}{An optional numeric value specifying the tolerance level for the iterative algorithm. Defaults to \code{1e-7}.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the \insertCite{aban2001;textual}{TIE} shifted Hill’s estimator for the tail index of a heavy-tailed distribution.
}
\details{
The \insertCite{aban2001;textual}{TIE} shifted Hill’s estimator adjusts the Hill estimator by introducing a shift parameter to better handle small sample sizes and reduce bias. The iterative algorithm continues until the difference between the maximum and minimum shift parameter estimates is below \code{eps}.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
am(x)
}
\references{
\insertRef{aban2001}{TIE}
}
