% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{AM1}
\alias{AM1}
\title{Aban & Meerschaert Hill Estimator}
\usage{
AM1(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the Aban & Meerschaert (2001) adjusted Hill estimator for the tail index of a heavy-tailed distribution.
}
\details{
The Aban & Meerschaert Hill estimator adjusts the standard Hill estimator by multiplying it by \code{k / (k - 1)} to correct for bias. This estimator is especially useful for small sample sizes where the bias in the Hill estimator is significant.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
AM1(x)

}
