% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{AM2}
\alias{AM2}
\title{Aban & Meerschaert Simplified Estimator}
\usage{
AM2(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the Aban & Meerschaert simplified estimator for the tail index of a heavy-tailed distribution.
}
\details{
The simplified Aban & Meerschaert estimator calculates the tail index using a weighted average of logarithmic values of the top-order statistics. The weights are derived from an adjusted cumulative sum of ranks. This estimator aims to provide a more straightforward alternative to the Hill estimator, incorporating a bias-correction mechanism.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
AM2(x)

}
