% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{AM3}
\alias{AM3}
\title{Aban & Meerschaert Smoothing Regression Estimator}
\usage{
AM3(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the Aban & Meerschaert smoothing regression estimator for the tail index of a heavy-tailed distribution.
}
\details{
The Aban & Meerschaert smoothing regression estimator uses a linear regression approach. The x-values for the regression are the negative logarithms of adjusted empirical ranks, and the y-values are the logarithms of the corresponding order statistics. The slope of the fitted regression line is used to calculate the tail index.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
AM3(x)

}
