% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{bb}
\alias{bb}
\title{\insertCite{bacro1995;textual}{TIE} estimator}
\usage{
bb(X, k, alpha)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}

\item{alpha}{An optional numeric value between 0 and 1 specifying the adjustment factor. Defaults to 0.9.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the \insertCite{bacro1995;textual}{TIE} estimator for the tail index of a heavy-tailed distribution.
}
\details{
\insertCite{bacro1995;textual}{TIE} estimator uses a refined \code{alpha} value to calculate the tail index, providing an alternative to other estimators like the Hill estimator.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
bb(x)
}
\references{
\insertRef{bacro1995}{TIE}
}
