% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{BCF3}
\alias{BCF3}
\title{Brito, Cavalcante, and Freitas (BCF) Adjusted Estimator (Exponential Correction)}
\usage{
BCF3(X, k, k2)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use for the initial BCF estimation. Defaults to the square root of the sample size.}

\item{k2}{An optional integer specifying the number of top-order statistics to use for the adjustment. Defaults to \(\max(0.9 \cdot n, k)\).}
}
\value{
A numeric value representing the adjusted tail index estimate.
}
\description{
Computes the adjusted Brito, Cavalcante, and Freitas (2016) estimator for the tail index of a heavy-tailed distribution using the method described in Equation 13 of their paper, with an exponential correction.
}
\details{
This estimator adjusts the BCF tail index by incorporating an exponential correction:
- The initial tail index estimate \eqn{g} is computed using the \code{BCF} function.
- The adjustment parameters (\eqn{\beta, \rho}) are computed using the \code{beta} and \code{rho} functions, respectively.
- The adjusted tail index (\eqn{g_2}) is calculated using the formula:
  \[
  g_2 = g \cdot \exp\left(-\frac{(n / k)^r \cdot \beta}{(1 - r)^2}\right),
  \]
  where \eqn{g} is the initial BCF estimate, (\eqn{\beta}) and (\eqn{\rho}) are adjustment parameters, and \( n \) and \( k \) are sample and group sizes.

This method provides a refined, bias-corrected estimate of the tail index using an exponential adjustment.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
BCF3(x)

}
