% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{BDG}
\alias{BDG}
\title{Beirlant, Dierckx, and Guillou (BDG) Estimator}
\usage{
BDG(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the BDG (2005) estimator for the tail index of a heavy-tailed distribution.
}
\details{
The BDG estimator calculates the tail index using the excess functions \code{UH}, derived from the order statistics and the Hill estimator. It involves a weighted mean of the logarithms of the excess functions, with weights based on the deviation of the empirical log-rank values from their mean.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
BDG(x)

}
