% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{BDG_H}
\alias{BDG_H}
\title{Beirlant, Dierckx, and Guillou (BDG) Estimator - Hill Modification}
\usage{
BDG_H(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the BDG (2005) estimator with the Hill modification for the tail index of a heavy-tailed distribution.
}
\details{
The BDG Hill modification uses the excess functions \code{UH}, derived from the order statistics and the Hill estimator, as input to the Hill estimator. This method refines the original BDG estimator by further reducing bias.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
BDG_H(x)

}
