% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{BF1}
\alias{BF1}
\title{Brito and Freitas Geometric Mean Estimator}
\usage{
BF1(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the Brito and Freitas estimator for the tail index of a heavy-tailed distribution. This estimator is the geometric mean of two estimators: the Zipf estimator and the reversed Schultze and Steinebach estimator.
}
\details{
The Brito and Freitas estimator combines the Zipf and the reversed Schultze and Steinebach estimators by taking their geometric mean. This approach leverages the strengths of both methods for more robust tail index estimation.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
BF1(x)

}
