% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{BFGV}
\alias{BFGV}
\title{Beirlant, Figueiredo, Gomes, and Vandewalle (2008) Estimator}
\usage{
BFGV(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the improved reduced-bias tail index estimator from Beirlant, Figueiredo, Gomes, and Vandewalle (2008).
}
\details{
This estimator refines the tail index by combining reduced-bias methods with auxiliary parameter adjustments:
- The auxiliary parameter \(\rho\) is computed using the \code{rho} function with \(k_1 = n^{0.9}\) order statistics.
- Several quantities are calculated for bias reduction:
  - \( D_0, D_1, D_2 \): Weighted sums of log differences for \(k_1\) top-order statistics.
  - \(\beta\): Bias adjustment parameter calculated from the weighted sums.
- The optimal number of order statistics \(k_0\) is determined as:
  \[
  k_0 = \left( \frac{(1 - 2r) \cdot n^{-2r}}{-2r \cdot \beta^2} \right)^{1/(1-2r)}.
  \]
  \( k_0 \) is constrained to be within \([k/10, k_1]\).
- The tail index is calculated as:
  \[
  \text{Tail Index} = \left[ H - \beta \cdot \left(\frac{n}{k}\right)^r \cdot D_0 \right]^{-1},
  \]
  where \( H \) is the Hill estimator.

This method reduces bias in the tail index estimation by leveraging auxiliary parameters and optimal order statistic selection.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
BFGV(x)

}
