% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tie.R
\name{bow}
\alias{bow}
\alias{neck}
\title{Classy neckwears}
\usage{
bow(data, x)

neck(data, x)
}
\description{
`bow` and `neck` can be combined to `tie` to do something similar to `dplyr::summarise` and `dplyr::mutate`, but allowing to express how to redistribute parts of the results to their own column.
}
\examples{
library(dplyr)

iris \%>\%
  group_by(Species) \%>\%
  bow( tie(min, max) := range(Sepal.Length) )

iris \%>\%
  group_by(Species) \%>\%
  neck( tie(min, max) := range(Sepal.Length) )

iris \%>\%
  group_by(Species) \%>\%
  bow( tie("min", "max") := range(Sepal.Length) )

small <- "min"
big   <- "max"
iris \%>\%
  group_by(Species) \%>\%
  bow( tie(!!small, !!big) := range(Sepal.Length) )

iris \%>\%
  group_by(Species) \%>\%
  bow( tie(,q25,q50,q75,) := quantile(Sepal.Length) )



}
