% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{BP}
\alias{BP}
\title{Baek and Pipiras (2010) Tail Index Estimator}
\usage{
BP(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the tail index using the Baek and Pipiras (2010) estimator, which is based on a regression model incorporating logarithmic and inverse transformations of order statistics.
}
\details{
This estimator applies a linear regression model to transformed order statistics to estimate the tail index:
- The top \( k \) order statistics are selected for the calculation, where \( k \) defaults to \( \sqrt{n} \).
- The regression model is constructed using:
  - \( y \): The logarithm of scaled ranks:
    \[
    y = \log\left(\frac{i - 0.5}{k}\right), \, i = 1, \ldots, k.
    \]
  - \( X_1 \): The negative logarithm of scaled order statistics:
    \[
    X_1 = -\log\left(\frac{X_{(n-i+1)}}{X_{(n-k)}}\right).
    \]
  - \( X_2 \): The inverse of scaled order statistics:
    \[
    X_2 = \frac{X_{(n-k)}}{X_{(n-i+1)}}, \, i = 1, \ldots, k.
    \]
- A linear regression model is fitted:
  \[
  y = \beta_1 \cdot X_1 + \beta_2 \cdot X_2 + \varepsilon,
  \]
  and the tail index is estimated as \( \beta_1 \).

This method incorporates both logarithmic and inverse transformations to improve the accuracy of the tail index estimation.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
BP(x)

}
