% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{BP3}
\alias{BP3}
\title{Baek and Pipiras (2010) Tail Index Estimator (Equation 2.18)}
\usage{
BP3(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the tail index using the Baek and Pipiras (2010) estimator based on their equation 2.18, which incorporates the Hill estimator and second moments.
}
\details{
This estimator combines the Hill estimator and the second moment of logarithmic differences to compute the tail index:
- The Hill estimator is calculated as:
  \[
  H = \text{Hill}(X, k).
  \]
- The second moment of logarithmic differences is computed as:
  \[
  M_2 = \text{MM}(X, k, 2),
  \]
  where \(\text{MM}(X, k, \alpha)\) computes the mean of the \(\alpha\)-th power of the logarithmic differences.
- The tail index is then calculated as:
  \[
  \text{Tail Index} = \left[\frac{(2 + 1/H) \cdot 2}{M_2} - 2 \cdot H \cdot \sqrt{\frac{2}{M_2}}\right].
  \]

This method leverages higher-order moments and the Hill estimator to enhance the accuracy of tail index estimation.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
BP3(x)

}
