% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{BVT}
\alias{BVT}
\title{Beirlant, Vynckier, and Teugels (BVT) Estimator}
\usage{
BVT(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the BVT estimator for the tail index of a heavy-tailed distribution using excess functions.
}
\details{
The BVT estimator calculates the tail index using the excess functions \code{UH}, which depend on the order statistics and the Hill estimator. A linear regression is performed on the logarithms of the excess functions and the negative logarithms of the empirical probabilities. The reciprocal of the slope of the regression line is the estimated tail index.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
BVT(x)

}
