% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{BVT2}
\alias{BVT2}
\title{Beirlant, Vynckier, and Teugels (BVT) Estimator with Anchor Point}
\usage{
BVT2(X, k, w)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}

\item{w}{An optional vector of weights for the regression. Defaults to equal weights for all points.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the BVT (1996) estimator for the tail index of a heavy-tailed distribution, using regression through the anchor point.
}
\details{
The BVT estimator with anchor point performs a weighted linear regression through the origin. The x-values are adjusted log-ranks, and the y-values are adjusted logarithms of the top-order statistics. The slope of the regression line is used to compute the tail index as its reciprocal.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
BVT2(x)

}
