% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{CG}
\alias{CG}
\title{Caeiro and Gomes (2002) Estimator}
\usage{
CG(X, k, alpha)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}

\item{alpha}{An optional numeric parameter for the moment order. If not provided, it is estimated from the auxiliary function \(\rho\). Defaults to \code{NULL}.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the tail index estimator from Caeiro and Gomes (2002).
}
\details{
This estimator combines moments with the auxiliary parameter \(\alpha\) to refine the tail index:
- If \(\alpha\) is not provided, it is estimated as:
  \[
  \alpha = \max\left(1, \frac{-\log(1 - r - \sqrt{(1 - r)^2 - 1})}{\log(1 - r)}\right),
  \]
  where \( r \) is computed using the \code{rho} function.
- The tail index is then calculated using:
  \[
  \text{Tail Index} = \frac{1}{\left( \frac{\Gamma(\alpha)}{\text{MM}(X, k, \alpha - 1)} \right) \cdot \sqrt{\frac{\text{MM}(X, k, 2\alpha)}{\Gamma(2\alpha + 1)}}}.
  \]

This method combines higher-order moments with auxiliary parameters to improve the accuracy of the tail index estimation.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
CG(x)

}
