% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{CGP2}
\alias{CGP2}
\title{Caeiro et al. (2005) Second Estimator}
\usage{
CGP2(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the second estimator from Caeiro et al. (2005) for the tail index of a heavy-tailed distribution.
}
\details{
This estimator refines the Hill estimator by incorporating exponential adjustments based on auxiliary functions:
- The auxiliary function \(\rho_1\) is computed using the \code{rho} function with \(\tau = 0\).
- The auxiliary function \(\beta_1\) is computed using the \code{beta} function.
- The first moment of the logarithmic differences of the top-order statistics (\( H \)) is computed using the \code{MM} function.
- The tail index is calculated using:
  \[
  \text{Tail Index} = \left[ H \cdot \exp\left(-\frac{\beta_1}{1 - \rho_1} \cdot \left( \frac{n}{k} \right)^{\rho_1}\right) \right]^{-1}.
  \]

This estimator combines the Hill estimator with exponential adjustments based on auxiliary functions for increased robustness.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
CGP2(x)

}
