% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{dHP}
\alias{dHP}
\title{De Haan and Pereira Estimator}
\usage{
dHP(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the De Haan and Pereira estimator for the tail index of a heavy-tailed distribution, as described in Equation (2) on page 40 of their paper.
}
\details{
This estimator is based on the ratio of the sum of squares of the top \( k \) order statistics to the sum of squares of the remaining order statistics:
- The top-order statistics are used to compute \( D_1 \), which is the product of \( k \) and the square of the \( (n-k) \)-th order statistic.
- The lower-order statistics are used to compute \( D_2 \), which is the sum of the squares of the remaining observations.
- The intermediate parameter (\eqn{\beta}) is defined as \( \beta = D_1 / D_2 \).
- The tail index is derived as \( \frac{2}{\beta + 1} \).
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
dHP(x)

}
