% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{dhr}
\alias{dhr}
\title{\insertCite{Haan1980;textual}{TIE} estimator}
\usage{
dhr(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the \insertCite{Haan1980;textual}{TIE} estimator for the tail index of a heavy-tailed distribution.
}
\details{
\insertCite{Haan1980;textual}{TIE} estimator is based on the logarithmic differences of the top-order statistics. It provides an alternative to the Hill estimator for tail index estimation.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
dhr(x)
}
\references{
\insertRef{Haan1980}{TIE}
}
