% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{DPR1}
\alias{DPR1}
\title{DPR Estimator}
\usage{
DPR1(X, m)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{m}{An optional integer specifying the group size for dividing the sample. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the DPR estimator for the tail index of a heavy-tailed distribution.
}
\details{
The DPR estimator divides the sorted sample into groups of size \code{m}. For each group, the ratio of the second-to-last order statistic to the maximum order statistic is calculated. The last group includes the remaining observations. The final estimate is computed based on these ratios and their sum.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
DPR1(x)

}
