% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{Falk}
\alias{Falk}
\title{Falk Estimator}
\usage{
Falk(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the Falk (1994) estimator for the tail index of a heavy-tailed distribution.
}
\details{
The Falk estimator is a weighted combination of two Pickands estimators: one using \code{k/2} order statistics and the other using \code{k} order statistics. The weights are set to \code{5/13} and \code{8/13}, respectively, to balance the contribution of the two estimators.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
Falk(x)

}
