% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{FH_MLE}
\alias{FH_MLE}
\title{Feuerverger and Hall (1999) Maximum Likelihood Tail Index Estimator}
\usage{
FH_MLE(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the tail index estimator using the Maximum Likelihood Estimator (MLE) approach proposed by Feuerverger and Hall (1999).
}
\details{
This function optimizes the MLE parameters (\code{D} and \(\rho\)) using the \code{FH_MLE1} function and applies the estimated parameters to compute the tail index:
- The parameters \code{D} and \(\rho\) are estimated by minimizing the negative log-likelihood via \code{optim}.
  - The optimization starts with initial values of \code{D = 1} and \(\rho = 1\).
  - The optimization is performed using the "L-BFGS-B" method with constraints: \(\rho \geq 0\).
- Using the optimized parameters, the tail index is calculated as:
  \[
  \text{Tail Index} = \left[ \frac{1}{k} \sum_{i=1}^{k} U_i \cdot \exp\left(-D \cdot \text{ind}_i\right)^{-\rho} \right]^{-1},
  \]
  where:
  \[
  U = i \cdot \left( \log(X_{(n-i+1)}) - \log(X_{(n-i)}) \right), \quad \text{ind}_i = \frac{i}{n}.
  \]

This method combines bias reduction through parameterized weighting and maximum likelihood estimation.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
FH_MLE(x)

}
