% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{fjp}
\alias{fjp}
\title{\insertCite{fialova2004;textual}{TIE} Estimator}
\usage{
fjp(X, N, alpha_0, delta)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{N}{An optional integer specifying the number of partitions. Defaults to 5.}

\item{alpha_0}{An optional initial estimate of the tail index. Defaults to 1.2 times the Hill estimator.}

\item{delta}{An optional numeric parameter controlling the adjustment of \code{a_N}. Defaults to 0.9.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the \insertCite{fialova2004;textual}{TIE} estimator for the tail index of a heavy-tailed distribution.
}
\details{
The \insertCite{fialova2004;textual}{TIE} estimator divides the sample into partitions, computes means for each partition, and calculates the tail index using an adjusted value of \code{a_N} and the empirical proportion of values below \code{a_N}. If the proportion \code{F} is not in the range (0, 1), the initial estimate \code{alpha_0} is returned.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
fjp(x)
}
\references{
\insertRef{fialova2004}{TIE}
}
