% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{GDPR}
\alias{GDPR}
\title{Generalized DPR (GDPR) Estimator with Permutations}
\usage{
GDPR(X, M)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{M}{An optional integer specifying the number of permutations to perform. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the Generalized DPR (GDPR) estimator for the tail index of a heavy-tailed distribution using multiple permutations of the data.
}
\details{
The GDPR estimator applies the \code{GDPR1} function to multiple permutations of the input data. 
- For each permutation, the tail index is computed using the \code{GDPR1} method, which supports generalized transformations.
- The final estimate is obtained as the mean of the tail index estimates across all permutations.
This approach enhances robustness to variations in the data by averaging estimates derived from randomized orders of the sample.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
GDPR(x)

}
