% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{GDPR1}
\alias{GDPR1}
\title{Generalized DPR (GDPR) Estimator}
\usage{
GDPR1(X, m, r)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{m}{An optional integer specifying the group size for dividing the sample. Defaults to the square root of the sample size.}

\item{r}{An optional numeric parameter controlling the transformation. Defaults to \code{-0.1}. When \code{r = 0}, the logarithmic transformation is used.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the Generalized DPR (GDPR) estimator for the tail index of a heavy-tailed distribution.
}
\details{
The GDPR estimator generalizes the DPR1 estimator by introducing a parameter \code{r} to control the transformation applied to the group-based ratios (\code{SS}). 
- The input data is divided into groups of size \code{m}, with the last group formed from the remaining observations.
- For each group, the ratio of the second-to-last order statistic to the maximum order statistic is calculated.
- Depending on the value of \code{r}, either a logarithmic transformation (\code{-log(SS)}) or a generalized transformation (\code{(1 - SS^r) / r}) is applied.
- The final tail index is computed based on the mean of the transformed values.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
GDPR1(x)

}
