% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{GFM}
\alias{GFM}
\title{Gomes et al. (2005) Best Linear Estimator}
\usage{
GFM(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the best linear estimator for the tail index from Gomes et al. (2005), as described on page 7 of their paper with \(\rho = -1\).
}
\details{
This estimator uses a weighted combination of Hill estimators to calculate the tail index:
- For each \( i \) from 1 to \( k \), the inverse Hill estimator (\( H_i \)) is computed for the top \( i \) order statistics.
- The weighted sum of the Hill estimators is calculated as:
  \[
  \text{Tail Index} = \left[ \frac{6}{k^2 - 1} \sum_{i=1}^{k-1} H_i \cdot i - H_k \cdot \frac{2k - 1}{k + 1} \right]^{-1}.
  \]

This method improves upon traditional Hill estimators by leveraging a linear weighting scheme for reduced bias, assuming \(\rho = -1\).
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
GFM(x)

}
