% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{GHR_port}
\alias{GHR_port}
\title{Gomes and Henriques-Rodrigues (2016) Tail Index Estimator with PORT Modification}
\usage{
GHR_port(X, k, q)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}

\item{q}{An optional numeric parameter controlling the PORT modification threshold. Defaults to \code{0.001}.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the tail index estimator from Gomes and Henriques-Rodrigues (2016), using a PORT modification of the Caeiro et al. (2005) estimator.
}
\details{
This estimator modifies the Caeiro et al. (2005) \code{CGP1} estimator by incorporating a PORT modification:
- The PORT modification shifts the data by removing the first \( kk = \text{round}(n \cdot q + 1) \) smallest elements and subtracting \( X_{(kk)} \):
  \[
  Y = X_{(kk+1):n} - X_{(kk)}.
  \]
- The \code{CGP1} estimator is then applied to the modified data \( Y \) using \( k \) as the number of top-order statistics.

This approach improves robustness and reduces bias by adapting the \code{CGP1} estimator to better handle the tail of the distribution.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
GHR_port(x)

}
