% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{GI1}
\alias{GI1}
\title{Gabaix & Ibragimov Reversed Estimator}
\usage{
GI1(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the Gabaix & Ibragimov reversed estimator for the tail index of a heavy-tailed distribution.
}
\details{
The Gabaix & Ibragimov reversed estimator applies a linear regression approach where the x-values are the logarithms of the order statistics and the y-values are the negative logarithms of adjusted empirical ranks. The slope of the fitted regression line is the tail index.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
GI1(x)

}
