% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{GI3}
\alias{GI3}
\title{Gabaix & Ibragimov Harmonic Estimator (Method 2)}
\usage{
GI3(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the Gabaix & Ibragimov Harmonic (Method 2) estimator for the tail index of a heavy-tailed distribution.
}
\details{
The Gabaix & Ibragimov Harmonic estimator (Method 2) reverses the linear regression setup by using the harmonic sums of ranks as the dependent variable (\code{H}) and the logarithms of the order statistics as the independent variable (\code{y}). The slope of the regression line is the estimated tail index.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
GI3(x)

}
