% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{GM1}
\alias{GM1}
\title{Gomes and Martin (2001) Estimator - Type 1}
\usage{
GM1(X, k, alpha)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}

\item{alpha}{An optional numeric parameter controlling the moment order. Defaults to \code{0.5}.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the Gomes and Martin (2001) Type 1 estimator for the tail index of a heavy-tailed distribution.
}
\details{
The Gomes and Martin Type 1 estimator computes the tail index using a combination of moments:
- The \( \alpha \)-th moment (\( M \)) and the first moment (\( H \)) of the logarithmic differences of the top-order statistics are computed using the \code{MM} function.
- The tail index is calculated as:
  \[
  \text{Tail Index} = \frac{\Gamma(\alpha + 1) \cdot H^{\alpha - 1}}{M},
  \]
  where \( \Gamma \) is the Gamma function, \( M \) is the \( \alpha \)-th moment, and \( H \) is the first moment.

This estimator incorporates a flexible parameter \( \alpha \), allowing for tailored moment-based calculations.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
GM1(x)

}
