% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{GM3}
\alias{GM3}
\title{Gomes and Martin (2002) Estimator - Type 1}
\usage{
GM3(X, k, tau)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}

\item{tau}{An optional numeric parameter for auxiliary adjustment. Defaults to \code{0}.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the first estimator from Gomes and Martin (2002) for the tail index of a heavy-tailed distribution.
}
\details{
This estimator refines the tail index calculation by incorporating the auxiliary function \(\rho_1\), the DJV estimator, and the Gomes-Martin Type 2 estimator:
- The auxiliary function \(\rho_1\) is computed using the \code{rho} function with \(\tau = 0\).
- The inverse of the DJV estimator (\( g_2 \)) is computed using the \code{DJV} function.
- The inverse of the Gomes-Martin Type 2 estimator (\( g_3 \)) is computed using the \code{GM2} function with \code{alpha = 2}.
- The tail index is calculated as:
  \[
  \text{Tail Index} = \frac{\rho_1}{-(2 - \rho_1) \cdot g_2 + 2 \cdot g_3}.
  \]

This estimator uses the auxiliary function \(\rho_1\) to account for additional structure in the data, improving bias and variance properties.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
GM3(x)

}
