% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{GM_FH}
\alias{GM_FH}
\title{Gomes and Martin (2004) Variant of Feuerverger and Hall MLE Estimator}
\usage{
GM_FH(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the tail index estimator using the Gomes and Martin (2004) variant of the Feuerverger and Hall Maximum Likelihood Estimator (MLE).
}
\details{
This estimator refines the Feuerverger and Hall MLE by incorporating additional moment calculations:
- Intermediate quantities are computed as follows:
  - \( U \): The scaled log differences:
    \[
    U = i \cdot \left(\log(X_{(n-i+1)}) - \log(X_{(n-i)})\right), \, i = 1, \ldots, n-1.
    \]
  - \( g_1 \): The Hill estimator for the tail index:
    \[
    g_1 = \frac{1}{\text{Hill}(X, k)}.
    \]
  - \( g_2 \): The mean of \( U \) over the top \( k \) order statistics:
    \[
    g_2 = \text{mean}(U_{1:k}).
    \]
  - \( g_3 \): The weighted sum of \( U \):
    \[
    g_3 = \sum_{i=1}^k \left(2i - k - 1\right) \cdot U_i.
    \]
  - \( g_4 \): The doubly weighted sum of \( U \):
    \[
    g_4 = \sum_{i=1}^k i \cdot \left(2i - k - 1\right) \cdot U_i.
    \]
- The tail index is calculated as:
  \[
  \text{Tail Index} = \left(g_1 - g_2 \cdot \frac{g_3}{g_4}\right)^{-1}.
  \]

This method enhances the original MLE by incorporating additional weighting terms to reduce bias.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
GM_FH(x)

}
