% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{GM_FH2}
\alias{GM_FH2}
\title{Gomes and Martin (2004) Variant of Feuerverger and Hall OLS Estimator}
\usage{
GM_FH2(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the tail index estimator using the Ordinary Least Squares (OLS) approach proposed by Gomes and Martin (2004), based on the Feuerverger and Hall methodology.
}
\details{
This estimator refines the Feuerverger and Hall OLS method by incorporating weighting and logarithmic transformations:
- Intermediate quantities are computed as follows:
  - \( U \): The scaled log differences:
    \[
    U = i \cdot \left(\log(X_{(n-i+1)}) - \log(X_{(n-i)})\right), \, i = 1, \ldots, n-1.
    \]
  - \( g_1 \): A weighted sum of log-transformed \( U \):
    \[
    g_1 = \frac{2(2k+1)}{k(k-1)} \cdot \sum_{i=1}^k \log(U_i).
    \]
  - \( g_2 \): Another weighted sum of log-transformed \( U \):
    \[
    g_2 = \frac{6}{k(k-1)} \cdot \sum_{i=1}^k i \cdot \log(U_i).
    \]
- The tail index is calculated as:
  \[
  \text{Tail Index} = \left[\exp\left(g_1 - g_2 + \gamma\right)\right]^{-1},
  \]
  where \(\gamma \approx 0.5772157\) is the Euler-Mascheroni constant.

This method enhances the original OLS approach by introducing additional weights to reduce bias and improve estimation accuracy.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
GM_FH2(x)

}
