% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{GMN1}
\alias{GMN1}
\title{Gomes et al. (2000) Jackknife Estimator - Type 1}
\usage{
GMN1(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the first Jackknife estimator from Gomes et al. (2000) for the tail index of a heavy-tailed distribution.
}
\details{
This Jackknife estimator combines the Hill estimator and the Danielsson, Jansen, and De Vries (DJV) estimator:
- The inverse of the Hill estimator (\( g_1 \)) is computed using the \code{Hill} function.
- The inverse of the DJV estimator (\( g_2 \)) is computed using the \code{DJV} function.
- The tail index is calculated as:
  \[
  \text{Tail Index} = \left[ 2 \cdot g_2 - g_1 \right]^{-1}.
  \]

This method uses a Jackknife approach to reduce bias by combining two different estimators.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
GMN1(x)

}
