% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{GMN3}
\alias{GMN3}
\title{Gomes et al. (2000) Jackknife Estimator - Type 3}
\usage{
GMN3(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the third Jackknife estimator from Gomes et al. (2000) for the tail index of a heavy-tailed distribution.
}
\details{
This Jackknife estimator combines the Danielsson, Jansen, and De Vries (DJV) estimator and the Gomes-Martin Type 2 estimator:
- The inverse of the DJV estimator (\( g_2 \)) is computed using the \code{DJV} function.
- The inverse of the Gomes-Martin Type 2 estimator (\( g_3 \)) is computed using the \code{GM2} function with \code{alpha = 2}.
- The tail index is calculated as:
  \[
  \text{Tail Index} = \left[ 3 \cdot g_2 - 2 \cdot g_3 \right]^{-1}.
  \]

This method uses a Jackknife approach to reduce bias by combining the DJV and Gomes-Martin estimators.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
GMN3(x)

}
