% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{GMN4}
\alias{GMN4}
\title{Gomes et al. (2000) Jackknife Estimator - Type 4}
\usage{
GMN4(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the fourth Jackknife estimator from Gomes et al. (2000) for the tail index of a heavy-tailed distribution.
}
\details{
This Jackknife estimator combines two Hill estimators using different subsets of the top-order statistics:
- The inverse of the Hill estimator (\( g_1 \)) is computed using \code{k / 2} order statistics.
- The inverse of the Hill estimator (\( g_2 \)) is computed using \( k \) order statistics.
- The tail index is calculated as:
  \[
  \text{Tail Index} = \left[ 2 \cdot g_1 - g_2 \right]^{-1}.
  \]

This method uses a Jackknife approach to reduce bias by combining Hill estimators computed from two different subsets of the data.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
GMN4(x)

}
