% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{GMNj2}
\alias{GMNj2}
\title{Gomes et al. (2002) Jackknife Estimator - Type 2}
\usage{
GMNj2(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the largest even integer less than or equal to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the second Jackknife estimator from Gomes et al. (2002) for the tail index of a heavy-tailed distribution.
}
\details{
This estimator refines the tail index using a Jackknife approach and two Hill estimators:
- The inverse Hill estimator (\( g_1 \)) is computed using \( k \) order statistics.
- The inverse Hill estimator (\( g_2 \)) is computed using \( k/2 \) order statistics.
- The weight \( w \) is calculated as:
  \[
  w = \frac{\log(1 - k / n)}{\log(1 - k / (2 \cdot n))}.
  \]
- The tail index is then computed as:
  \[
  \text{Tail Index} = \frac{1 - w}{g_1 - g_2 \cdot w}.
  \]

This method reduces bias by combining two Hill estimators with a weighting factor \( w \).
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
GMNj2(x)

}
