% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{GMNj3}
\alias{GMNj3}
\title{Gomes et al. (2002) Jackknife Estimator - Type 3}
\usage{
GMNj3(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the largest even integer less than or equal to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the third Jackknife estimator from Gomes et al. (2002) for the tail index of a heavy-tailed distribution.
}
\details{
This estimator refines the tail index using a Jackknife approach and two Hill estimators:
- The inverse Hill estimator (\( g_1 \)) is computed using \( k \) order statistics.
- The inverse Hill estimator (\( g_2 \)) is computed using \( k/2 \) order statistics.
- The tail index is then computed as:
  \[
  \text{Tail Index} = \frac{1 + k / n}{g_2 \cdot (2 + k / n) - g_1}.
  \]

This method combines two Hill estimators with an adjustment term \( k / n \) to improve bias and variance properties.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
GMNj3(x)

}
