% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{GMV2}
\alias{GMV2}
\title{Gomes, Miranda, and Viseu (Statistica Neerlandica) Estimator - Type 2}
\usage{
GMV2(X, k, alpha)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}

\item{alpha}{An optional numeric parameter controlling the moment order. Defaults to \code{1.5}.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the second tail index estimator from Gomes, Miranda, and Viseu (Statistica Neerlandica, equation 6, page 3).
}
\details{
This estimator refines the tail index by combining a weighted sum of differences between log-transformed order statistics, incorporating logarithmic adjustments:
- The weights depend on the moment order parameter \(\alpha\), the rank of the order statistics, and a logarithmic scaling factor.
- The tail index is calculated as:
  \[
  \text{Tail Index} = \left[ -\frac{\alpha^2}{k} \cdot \sum_{i=1}^{k} \left( \frac{i}{k} \right)^{\alpha - 1} \cdot \log\left(\frac{i}{k}\right) \cdot i \cdot \left( \log(X_{(n-i+1)}) - \log(X_{(n-i)}) \right) \right]^{-1}.
  \]

This method uses the parameter \(\alpha\) to adjust the influence of order statistics, with \(\alpha = 1.5\) as a common default, and logarithmic scaling to further refine the weights.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
GMV2(x)

}
