% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{GMV5}
\alias{GMV5}
\title{Gomes, Miranda, and Viseu (Statistica Neerlandica) Estimator - Type 5 (\(\rho = -1\))}
\usage{
GMV5(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the tail index estimator from Gomes, Miranda, and Viseu (Statistica Neerlandica, page 15, \(\rho = -1\)).
}
\details{
This estimator refines the tail index by combining two generalized estimators:
- Two generalized estimators are computed:
  - \( g_1 \): The generalized estimator using \(\text{GMV1}\) with \(\alpha = 1\).
  - \( g_2 \): The generalized estimator using \(\text{GMV1}\) with \(\alpha = 2\).
- The tail index is calculated as:
  \[
  \text{Tail Index} = \left[ 4 \cdot g_1 - 3 \cdot g_2 \right]^{-1}.
  \]

This method reduces bias by combining the two estimators in a weighted linear fashion, assuming \(\rho = -1\).
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
GMV5(x)

}
