% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{GPM}
\alias{GPM}
\title{Gomes et al. (2005) Generalized Estimator}
\usage{
GPM(X, s, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{s}{An optional numeric parameter controlling the order of the estimator. Defaults to \code{2} (with \code{s >= 1}).}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the generalized tail index estimator from Gomes et al. (2005), as described on page 3 of their paper.
}
\details{
This estimator generalizes the tail index calculation by introducing a parameter \( s \) that adjusts the influence of the order statistics:
- The tail index is calculated using:
  \[
  \text{Tail Index} = \left[ \frac{s^2}{k^s} \cdot \sum_{i=1}^{k} i^{s-1} \cdot \log\left( \frac{X_{(n-k+i)}}{X_{(n-k)}} \right) \right]^{-1},
  \]
  where \( X_{(i)} \) represents the \( i \)-th order statistic of the sorted sample, \( k \) is the number of top-order statistics, and \( s \) is the order parameter.

This method allows for greater flexibility in estimating the tail index by adjusting the parameter \( s \), with \( s \geq 1 \).
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
GPM(x)

}
