% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{GPMj}
\alias{GPMj}
\title{Gomes et al. (2005) Jackknife Estimator}
\usage{
GPMj(X, s, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{s}{An optional numeric parameter controlling the order of the estimator. Defaults to \code{2} (with \code{s >= 1}).}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the Jackknife estimator from Gomes et al. (2005), as described on page 3 of their paper.
}
\details{
This estimator refines the generalized tail index estimator by using a Jackknife adjustment:
- The generalized estimators \( g_1 \) and \( g_2 \) are computed using \code{GPM} with order parameters \( 1 \) and \( s \), respectively.
- The auxiliary function \(\rho_1\) is computed using the \code{rho} function with \code{tau = 1}. It is recommended to use \(\rho_1 = -1\) for improved performance.
- The tail index is calculated as:
  \[
  \text{Tail Index} = \left[ -\frac{s \cdot (1 - \rho_1)}{\rho_1 \cdot (s - 1)} \cdot \left( g_1 - \frac{s - \rho_1}{s \cdot (1 - \rho_1)} \cdot g_2 \right) \right]^{-1}.
  \]

This method combines the generalized estimators with a Jackknife adjustment for bias reduction.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
GPMj(x)

}
