% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{GRM_port}
\alias{GRM_port}
\title{Gomes and Henriques-Rodrigues (2016) Tail Index Estimator with PORT Modification}
\usage{
GRM_port(X, k, q)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}

\item{q}{An optional numeric parameter controlling the PORT modification threshold. Defaults to \code{0.001}.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the tail index estimator from Gomes and Henriques-Rodrigues (2016, REVSTAT), incorporating a PORT modification.
}
\details{
This estimator refines the tail index by applying the PORT modification to the data before applying the \code{BGP} estimator:
- The PORT modification shifts the data by removing the first \( kk = \text{round}(n \cdot q + 1) \) smallest elements and subtracting \( X_{(kk)} \):
  \[
  Y = X_{(kk+1):n} - X_{(kk)}.
  \]
- The \code{BGP} estimator is then applied to the modified data \( Y \) using \( k \) as the number of top-order statistics.

This method improves robustness and reduces bias in the tail index estimation by modifying the data to better fit the tail distribution.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
GRM_port(x)

}
