% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{hill}
\alias{hill}
\title{\insertCite{hill1975;textual}{TIE} estimator}
\usage{
hill(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the \insertCite{hill1975;textual}{TIE} estimator, a commonly used method for estimating the tail index of heavy-tailed distributions.
}
\details{
\insertCite{hill1975;textual}{TIE} estimator is a semi-parametric estimator based on the order statistics of the data. It is used to estimate the tail index, which characterizes the heaviness of the tail of a distribution.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
hill(x)
}
\references{
\insertRef{hill1975}{TIE}
}
