% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{HWm}
\alias{HWm}
\title{Hosking-Wallis Tail Index Estimator (Method of Moments)}
\usage{
HWm(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the full sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the tail index using the Hosking-Wallis method of moments estimator, which is based on the mean and variance of the top-order statistics.
}
\details{
This estimator uses the mean and variance of the top-order statistics to compute the tail index:
- The top \( k \) order statistics are selected, where \( k \) defaults to the full sample size (\( n \)).
- The mean (\( \bar{x} \)) and variance (\( s^2 \)) of the selected order statistics are calculated:
  \[
  \bar{x} = \text{mean}(X_{(n-k+1):n}),
  \quad s^2 = \text{var}(X_{(n-k+1):n}).
  \]
- The tail index is then computed as:
  \[
  \text{Tail Index} = \left[0.5 \cdot \left(1 - \frac{\bar{x}^2}{s^2}\right)\right]^{-1}.
  \]

This method leverages the relationship between the mean and variance of extreme values to estimate the tail index.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
HWm(x)

}
