% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{JP2004}
\alias{JP2004}
\title{Jurečková and Picek Estimator with Permutations (2004)}
\usage{
JP2004(X, M)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{M}{An optional integer specifying the number of permutations to perform. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the Jurečková and Picek (2004) estimator for the tail index of a heavy-tailed distribution using multiple permutations of the data.
}
\details{
This estimator applies the \code{JP2004a} function to multiple permutations of the input data:
- For each permutation, the sample is divided into groups, and the maxima of the groups are used to estimate the tail index.
- The final estimate is obtained as the mean of the tail index estimates across all permutations.

Permutations enhance robustness by averaging the estimates derived from different random orders of the sample.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
JP2004(x)

}
