% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{JP2004a}
\alias{JP2004a}
\title{Jurečková and Picek Estimator (2004)}
\usage{
JP2004a(X, m, delta)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{m}{An optional integer specifying the group size for dividing the sample. Defaults to the square root of the sample size.}

\item{delta}{An optional numeric parameter controlling the scaling factor. Defaults to \code{0.05}.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the Jurečková and Picek (2004) estimator for the tail index of a heavy-tailed distribution.
}
\details{
This estimator divides the input data into groups of size \code{m}, computes the maximum of each group, and derives the tail index using a quantile-based scaling formula:
- The input data is divided into \code{T} groups, and the maximum value of each group is stored in \code{SS}.
- The scaling factors \code{a1} and \code{a2} are calculated based on the group size (\code{m}) and parameter \code{delta}.
- The tail index is derived using the logarithm of \code{a1} and the quantile of the maxima vector \code{SS}.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
JP2004a(x)

}
