% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{MEP_BAS}
\alias{MEP_BAS}
\title{McElroy and Politis BAS Estimator}
\usage{
MEP_BAS(X, r)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{r}{An optional positive integer specifying the moment order. Defaults to \code{1}. \code{r} should be relatively large, such that the \code{2r}-th moment does not exist.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the McElroy and Politis (2006) BAS estimator for the tail index of a heavy-tailed distribution.
}
\details{
The McElroy and Politis BAS estimator calculates the tail index based on the \code{2r}-th moment of the sample. The method involves a logarithmic transformation of the sum of \code{2r}-th powers of the data and the sample size.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
MEP_BAS(x)

}
