% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{MEP_CEN}
\alias{MEP_CEN}
\title{McElroy and Politis CEN Estimator}
\usage{
MEP_CEN(X)

MEP_CEN(X)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}
}
\value{
A numeric value representing the estimated tail index.

A numeric value representing the estimated tail index.
}
\description{
Computes the McElroy and Politis (2006) CEN estimator for the tail index of a heavy-tailed distribution.

Computes the McElroy and Politis (2006) CEN estimator for the tail index of a heavy-tailed distribution.
}
\details{
The McElroy and Politis CEN estimator computes the tail index based on the second moments of the entire sample and the top \code{k} observations, where \code{k} is the square root of the sample size. This method relies on the logarithmic difference of these second moments.

The McElroy and Politis CEN estimator computes the tail index based on the second moments of the entire sample and the top \code{k} observations, where \code{k} is the square root of the sample size. This method relies on the logarithmic difference of these second moments.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
MEP_CEN(x)

set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
MEP_CEN(x)

}
