% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{MEP_RCEN}
\alias{MEP_RCEN}
\title{McElroy and Politis RCEN Estimator}
\usage{
MEP_RCEN(X)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the McElroy and Politis (2006) RCEN (revised CEN) estimator for the tail index of a heavy-tailed distribution.
}
\details{
The RCEN estimator divides the sample into \code{b} blocks of size \code{sqrt(n)}, where \code{b} is the floor of the square root of the sample size. For each block, the estimator is calculated using the logarithmic difference of the second moments of the entire sample and the block. The final tail index is computed as the reciprocal of the mean of the block estimates.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
MEP_RCEN(x)

}
