% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{ms}
\alias{ms}
\title{\insertCite{meerschaert1998;textual}{TIE} estimator}
\usage{
ms(X)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the \insertCite{meerschaert1998;textual}{TIE} estimator for the tail index of a heavy-tailed distribution.
}
\details{
The \insertCite{meerschaert1998;textual}{TIE} estimator calculates the tail index using the variance of the sample around its mean. It is based on the relationship between the logarithm of the sample size and the logarithm of the sample variance.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
ms(x)
}
\references{
\insertRef{meerschaert1998}{TIE}
}
