% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{muller}
\alias{muller}
\title{\insertCite{muller2003;textual}{TIE} Estimator}
\usage{
muller(X, m, a1, a2)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{m}{An optional integer specifying the number of observations to use. Defaults to the square root of the sample size.}

\item{a1}{An optional numeric value representing the first threshold parameter. Defaults to one-fourth of the range of the top \code{m} observations.}

\item{a2}{An optional numeric value representing the second threshold parameter. Defaults to half of the range of the top \code{m} observations.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the \insertCite{muller2003;textual}{TIE} estimator for the tail index of a heavy-tailed distribution.
}
\details{
The \insertCite{muller2003;textual}{TIE} estimator calculates the tail index using two threshold parameters (\code{a1} and \code{a2}) and indicator functions applied to the data. The thresholds are adaptively determined based on the range of the top \code{m} observations if not provided.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
muller(x)
}
\references{
\insertRef{muller2003}{TIE}
}
