% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{mullersp}
\alias{mullersp}
\title{\insertCite{muller2009;textual}{TIE} Smoothed Pickands Estimator}
\usage{
mullersp(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the \insertCite{muller2009;textual}{TIE} smoothed Pickands estimator for the tail index of a heavy-tailed distribution.
}
\details{
The \insertCite{muller2009;textual}{TIE} smoothed Pickands estimator employs logspline density estimation to smooth the quantiles of the data before calculating the tail index. It uses three quantiles (\code{X1}, \code{X2}, and \code{X4}) derived from the logspline density estimate to refine the Pickands ratio calculation.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
mullersp(x)
}
\references{
\insertRef{muller2009}{TIE}
}
