% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{nuyts}
\alias{nuyts}
\title{\insertCite{nuyts2010;textual}{TIE} Estimator}
\usage{
nuyts(X, k, eps = 1e-09)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}

\item{eps}{An optional numeric value specifying the tolerance level for the iterative algorithm. Defaults to \code{1e-9}.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the \insertCite{nuyts2010;textual}{TIE} estimator for the tail index of a heavy-tailed distribution.
}
\details{
The \insertCite{nuyts2010;textual}{TIE} estimator uses a binary search algorithm to determine the tail index. The iterative process continues until the difference between the maximum and minimum estimates is below \code{eps}. If the difference exceeds a threshold, a second iteration is performed to refine the estimate.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
Nuyts(x)
}
\references{
\insertRef{nuyts2010}{TIE}
}
