% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{peng_h}
\alias{peng_h}
\title{\insertCite{peng1998;textual}{TIE} Estimator (Hill Modification)}
\usage{
peng_h(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the \insertCite{peng1998;textual}{TIE}, a modification of the Hill estimator, for estimating the tail index of a heavy-tailed distribution.
}
\details{
The \insertCite{peng1998;textual}{TIE} refines the Hill estimator by incorporating additional moments and parameters to reduce bias. It adjusts the tail index estimation using intermediate order statistics.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
peng_h(x)
}
\references{
\insertRef{peng1998}{TIE}
}
