% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{peng_p}
\alias{peng_p}
\title{\insertCite{peng1998;textual}{TIE} Estimator (Pickands Modification)}
\usage{
peng_p(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the \insertCite{peng1998;textual}{TIE} estimator, a modification of the Pickands estimator, for estimating the tail index of a heavy-tailed distribution.
}
\details{
The \insertCite{peng1998;textual}{TIE} estimator modifies the Pickands estimator by adjusting the tail index computation using intermediate order statistics and introducing a parameter \code{rho} for bias correction.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
peng_p(x)
}
\references{
\insertRef{peng1998}{TIE}
}
