% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{pickands}
\alias{pickands}
\title{Pickands estimator}
\usage{
pickands(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size. The value of \code{k} is adjusted to be divisible by 4.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the \insertCite{pickands1975;textual}{TIE} estimator for the tail index of a heavy-tailed distribution.
}
\details{
\insertCite{pickands1975;textual}{TIE} estimator is based on the ratios of spacings of extreme order statistics. It is a commonly used method for estimating the tail index of heavy-tailed distributions.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
pickands(x)
}
\references{
\insertRef{pickands1975}{TIE}
}
