% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{pl}
\alias{pl}
\title{\insertCite{politis2002;textual}{TIE} estimator}
\usage{
pl(X)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the \insertCite{politis2002;textual}{TIE} tail index estimator for a heavy-tailed distribution using a bootstrap approach.
}
\details{
The \insertCite{politis2002;textual}{TIE} estimator applies a bootstrap procedure to compute the median of tail index estimates derived from resampled datasets. The tail index for each resampled dataset is calculated using the \code{pl} function, which estimates the power-law tail index.

The function uses \code{m = 1000} bootstrap iterations by default.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
pl(x)
}
\references{
\insertRef{politis2002}{TIE}
}
