% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholdselection.R
\name{plotuniform}
\alias{plotuniform}
\title{QQ plot for fitted tail index and quantile}
\usage{
plotuniform(wn, tail, Y)
}
\arguments{
\item{wn}{quantile or threshold}

\item{tail}{tail index}

\item{Y}{data}
}
\value{
a figure cantians QQ plot
}
\description{
You can use this method to check whether you fit the data well.
}
\examples{
\dontrun{
p = 0.9
wn = qr(x, p) ##x is the data
tail = TIE.dynamic1(x,...)
plot.uniform(wn, tail, x)
}
}
