% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QR.R
\name{qr}
\alias{qr}
\title{Using CAViaR method to estimate high but not extreme quantile}
\usage{
qr(x, p, tol = 1e-04)
}
\arguments{
\item{x}{time series data}

\item{p}{the quantile you want to estimate}

\item{tol}{the accuracy of your optimization}
}
\value{
A vector containing the estimated beta
}
\description{
Using CAViaR method to estimate high but not extreme quantile
}
\examples{
\dontrun{
x = rgev(1000)
p = 0.9
qr(x, p)
qr(x, 0.85)

}
}
