% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recons.R
\name{recon.q}
\alias{recon.q}
\title{reconstruct quantile use the estimated beta}
\usage{
recon.q(beta, init, x)
}
\arguments{
\item{beta}{estimated beta}

\item{init}{the initial value for quantile regression}

\item{x}{data}
}
\value{
a vector that contains the quantile series
}
\description{
This function use the estimated beta to reconstruct the quantile
}
\examples{
\dontrun{
x ##data
p = 0.9
init = qnorm(p)
beta.es = qr(x, init)
recon.q(beta.es, p)
}
}
