% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recons.R
\name{recon.tail}
\alias{recon.tail}
\title{reconstruct tail using the estimated gamma}
\usage{
recon.tail(x, gam, real = TRUE, G = 10)
}
\arguments{
\item{x}{data}

\item{gam}{estimated gamma}

\item{real}{this identify wether you use the true data, or you use the simulation data.}

\item{G}{the value fitted into the dynamic equation}
}
\value{
A vector contains the estimated tail
}
\description{
This function use the estimated gamma to reconstruct the tail
}
\examples{
\dontrun{
x ##data
p = 0.9
q.es = qr(x, p) ##estimated quantile as threshold
gam.es = TIE.dynamic1(q.es, x)
tail.es = recon.tail(x, gam.es)
}
\dontrun{
##When doing simulation, I would call TIE.dynamic2, and let the real = FALSE
x ##data
p = 0.9
q.es = qr(x, p) ##estimated quantile as threshold
gam.es = TIE.dynamic2(q.es, x)
tail.es = recon.tail(x, gam.es, real = False)
}
}
\seealso{
\code{\link{recon.q}}
}
