% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{SAG}
\alias{SAG}
\title{SAG Tail Index Estimator}
\usage{
SAG(X, k, q)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size of the adjusted data.}

\item{q}{An optional numeric parameter controlling the threshold for the adjustment. Defaults to \code{0.1}.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the tail index using the SAG estimator, which incorporates a threshold adjustment and applies the Hill estimator on the adjusted data.
}
\details{
This estimator applies a threshold adjustment to the data before computing the tail index using the Hill estimator:
- The adjustment threshold is defined as \( m = \text{round}(n \cdot q + 1) \), where \( q \) controls the fraction of the sample used for adjustment.
- The adjusted data is calculated as:
  \[
  Y = X_{(m+1):n} - X_{(m)}.
  \]
- The Hill estimator is then applied to the adjusted data \( Y \), using \( k \) as the number of top-order statistics.

This approach ensures that the tail index estimation focuses on the extreme values of the distribution, while removing potential bias introduced by smaller values.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
SAG(x)

}
