% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{Smith}
\alias{Smith}
\title{Smith Tail Index Estimator}
\usage{
Smith(X, u)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{u}{An optional numeric value specifying the threshold. If not provided, the threshold is set as the \( (n - k) \)-th largest value, where \( k = \sqrt{n} \).}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the tail index using the Smith estimator, which applies a logarithmic transformation to exceedances above a threshold.
}
\details{
This estimator focuses on exceedances above a specified threshold to calculate the tail index:
- If the threshold \( u \) is not provided, it is set to:
  \[
  u = X_{(n-k)},
  \]
  where \( k = \sqrt{n} \).
- The exceedances \( Y \) above the threshold \( u \) are defined as:
  \[
  Y = X[X > u] - u.
  \]
- The transformed values \( Z \) are calculated as:
  \[
  Z = \log\left(\frac{Y}{u} + 1\right).
  \]
- The tail index is estimated as the inverse of the mean of \( Z \):
  \[
  \text{Tail Index} = \left(\text{mean}(Z)\right)^{-1}.
  \]

This method effectively captures the behavior of the tail by focusing on values exceeding a dynamically chosen threshold.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
Smith(x)

}
