% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{SS}
\alias{SS}
\title{Schultze and Steinebach Estimator}
\usage{
SS(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the Schultze and Steinebach (1996) estimator for the tail index of a heavy-tailed distribution.
}
\details{
The Schultze and Steinebach estimator uses a linear regression approach without an intercept. The logarithms of scaled ranks are used as the independent variable, and the logarithms of the top-order statistics are used as the dependent variable. The ratio of the sums of squared logarithms and the cross-product of the logarithms provides the tail index estimate.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
SS(x)

}
