% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thresholdselection.R
\name{statistics}
\alias{statistics}
\title{Threshold selection}
\usage{
statistics(wn, tail, Y)
}
\arguments{
\item{wn}{threshold quantile}

\item{tail}{tail index}

\item{Y}{data}
}
\value{
A value that indicate whether the threshold selected is good
}
\description{
This function will be used as threshold selection in my method, it returns a value that indicate
whether the threshold selected is good, using the criteria in the Tail Index Regression. It should
be used after the you get the quantile series and tail series.
And you can use the plot.uniform function to see the figure.
}
\examples{
\dontrun{
p = 0.9
wn = qr(x, p) ##x us the data
tail = TIE(x,...)
statistics(wn, tail, x)
}
}
\seealso{
\code{\link{plotuniform}}
}
