% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TIE.R
\name{TIE.dynamic1}
\alias{TIE.dynamic1}
\alias{TIE.dynamic2}
\alias{TIE.fix}
\title{Tail Index Estimation}
\usage{
TIE.dynamic1(q.es, x, init = 0.3)

TIE.dynamic2(x, q.es, G = 10, init = 0.3)

TIE.fix(x, q.es)
}
\arguments{
\item{q.es}{estimated quantile threshold}

\item{x}{data}

\item{init}{initial value for tail index}

\item{G}{You can see the article for the selection of G}
}
\value{
a list that contains the estimated gamma

a vector that contains the estimated gamma which will be used to fit into the recon.tail function

the estimated fixed tail index
}
\description{
These method combines the extreme value theory and CAViaR to obtain the dynamic of tail index
we first get some high quantile estimation in the qr function, and then use this as threshold
and use the extreme value theory to get a likelihood type criteria.The function returns the estimated
gamma and then you can use the estimated gamma and fit this in the recon.tail function to get the
tail index series.
}
\details{
Version one spend some time choosing the initial value of the optimization. If you find that
the TIE estimation version2 lacks the stability you can use this version, this version will
use different value for initiallization, so it can give more reliable solution.

I recommend to use version two in simulation, since the dynamic of the tail index is simpler
in order to increase the speed in simulation.

And TIE.fix assumes the tail index is a constant, which leads to quicker computation, but gives
less appealing result.
}
\section{Functions}{
\itemize{
\item \code{TIE.dynamic2}: Another version of TIE.dynamic

\item \code{TIE.fix}: Fixed TIE estimation
}}

\examples{
\dontrun{
x ##data
q.es = qr(x, 0.9)
TIE.dynamic1(q.es, x)
}
\dontrun{
x ##data
q.es = qr(x, 0.9)
TIE.dynamic1(q.es, x)
}
\dontrun{
x ##data
p = 0.9
q.es = qr(x, p)
tail.es = TIE.fix(x, q.es)
}
}
