% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{tkp2}
\alias{tkp2}
\title{\insertCite{tripathi2014;textual}{TIE} Type II Estimator}
\usage{
tkp2(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the \insertCite{tripathi2014;textual}{TIE} type II estimator for the tail index of a heavy-tailed distribution.
}
\details{
The \insertCite{tripathi2014;textual}{TIE} type II estimator is an extension of the TKP1 estimator that includes an additional term \code{max(0, log(X[n-k]))} in the denominator to improve robustness. It is based on equation 3.2 from the original paper.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
tkp2(x)
}
\references{
\insertRef{tripathi2014}{TIE}
}
