% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{tpk1}
\alias{tpk1}
\title{\insertCite{tripathi2014;textual}{TIE} Type I Estimator}
\usage{
tpk1(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the \insertCite{tripathi2014;textual}{TIE} type I estimator for the tail index of a heavy-tailed distribution.
}
\details{
The \insertCite{tripathi2014;textual}{TIE} type I estimator is based on the ratio of logarithmic differences of top-order statistics. It is designed to provide an efficient estimate of the tail index using equation 3.2 from the original paper.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
tpk1(x)
}
\references{
\insertRef{tripathi2014}{TIE}
}
