% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{Vaiciulis}
\alias{Vaiciulis}
\title{Vaiciulis Estimator (2012)}
\usage{
Vaiciulis(X, m, l)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{m}{An optional integer specifying the group size for dividing the sample. Defaults to the square root of the sample size.}

\item{l}{An optional integer specifying the maximum order of the power series. Defaults to 3.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the Vaiciulis (2012) estimator for the tail index of a heavy-tailed distribution.
}
\details{
The Vaiciulis estimator divides the input data into groups of size \code{m}. 
- Within each group, the logarithm of the ratio of the maximum order statistic to the second-to-last order statistic is computed and raised to successive powers up to \code{l+1}.
- The method calculates the averages of these transformed values across groups and applies a specific power series formula involving gamma functions to compute the tail index.
- The last group includes the remaining observations if the sample size is not divisible by \code{m}.

This estimator involves higher-order corrections to improve the accuracy of tail index estimation.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
Vaiciulis(x)

}
