% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{vanzyl}
\alias{vanzyl}
\title{\insertCite{zyl2015;textual}{TIE} Estimator}
\usage{
vanzyl(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the \insertCite{zyl2015;textual}{TIE} estimator for the tail index of a heavy-tailed distribution.
}
\details{
The \insertCite{zyl2015;textual}{TIE} estimator incorporates the Hill estimator along with location and scale parameters to compute the tail index. It uses an adjusted logarithmic mean of the top-order statistics for better accuracy.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
vanzyl(x)
}
\references{
\insertRef{zyl2015}{TIE}
}
