% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{vbch}
\alias{vbch}
\title{\insertCite{vandewalle2007;textual}{TIE} Estimator}
\usage{
vbch(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the \insertCite{vandewalle2007;textual}{TIE} estimator for the tail index of a heavy-tailed distribution.
}
\details{
The \insertCite{vandewalle2007;textual}{TIE} estimator minimizes the objective function defined by the \code{VBCH1} function using the \code{optim} function with the \code{L-BFGS-B} method. It uses starting values and constraints for the parameters (\code{gamma}, \code{delta}, \code{rho}, and \code{w}). The tail index is computed as the reciprocal of the optimized value of \code{gamma}.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
vbch(x)
}
\references{
\insertRef{vandewalle2007}{TIE}
}
