% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{VP_2}
\alias{VP_2}
\title{Vaiciulis and Paulauskas Gamma\(_2\) Estimator}
\usage{
VP_2(X, k, r)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}

\item{r}{An optional numeric parameter controlling the transformation. Defaults to \code{-0.1}.}
}
\value{
A numeric value representing the estimated tail index \eqn{\gamma_2}.
}
\description{
Computes the Vaiciulis and Paulauskas (2017) Gamma\(_2\) estimator (\eqn{\gamma_2}) for the tail index of a heavy-tailed distribution, as described on page 6 of their paper.
}
\details{
This estimator relies on the transformation provided by the Vaiciulis and Paulauskas \eqn{g} function:
- The function \eqn{g} is called with \( r \) and logarithmic transformation parameter \( u = 1 \).
- The tail index \eqn{\gamma_2} is computed using the formula:
  \[
  \gamma_2 = \frac{D_2}{D_1}, \quad \text{where } D_1 = 2 \cdot G_1 \text{ and } D_2 = 2r \cdot G_1 + 1 + \sqrt{4r \cdot G_1 + 1}.
  \]
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
VP_2(x)

}
