% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{VP_3}
\alias{VP_3}
\title{Vaiciulis and Paulauskas Gamma\(_3\) Estimator}
\usage{
VP_3(X, k, r)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}

\item{r}{An optional numeric parameter controlling the transformation. Defaults to \code{-0.1}.}
}
\value{
A numeric value representing the estimated tail index \(\gamma_3\).
}
\description{
Computes the Vaiciulis and Paulauskas (2017) Gamma\(_3\) estimator (\(\gamma_3\)) for the tail index of a heavy-tailed distribution, as described on page 6 of their paper.
}
\details{
This estimator uses the transformation provided by the Vaiciulis and Paulauskas \eqn{g} function:
- The function \eqn{g} is called with parameters \( r \) and \( u = 0 \) to compute \( G_0 \), and with \( u = 1 \) to compute \( G_1 \).
- The tail index \(\gamma_3\) is computed using the formula:
  \[
  \gamma_3 = \begin{cases} 
  \text{DJV}(X), & \text{if } r = 0, \\
  \frac{D_2}{D_1}, & \text{otherwise},
  \end{cases}
  \]
  where \( D_1 = r \cdot G_1 - G_0 + 1 \) and \( D_2 = r^2 \cdot G_1 \).

If \( r = 0 \), the estimator defaults to the Danielsson, Jansen, and De Vries (DJV) estimator.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
VP_3(x)

}
