% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{VP_M}
\alias{VP_M}
\title{Vaiciulis and Paulauskas Moment Estimator Modification}
\usage{
VP_M(X, k, r)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}

\item{r}{An optional numeric parameter controlling the transformation. Defaults to \code{-0.1}. When \code{r = 0}, the logarithmic transformation is used.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the Vaiciulis and Paulauskas (2013) modification of the moment estimator for the tail index of a heavy-tailed distribution.
}
\details{
The Vaiciulis and Paulauskas modification of the moment estimator introduces a parameter \code{r} for a generalized transformation of the scaled top-order statistics:
- When \code{r = 0}, the transformation is logarithmic (\code{log(Y)}).
- When \code{r != 0}, the transformation is \code{(Y^r - 1) / r}, where \code{Y} is the ratio of the top-order statistics to the threshold value.
- Higher-order moments (\code{M1, M12, M2}) are calculated and used to derive the tail index with a bias correction.

This modification enhances the flexibility and robustness of the moment estimator by accommodating different transformations based on the value of \code{r}.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
VP_M(x)

}
