% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{Weis}
\alias{Weis}
\title{Weis Tail Index Estimator}
\usage{
Weis(X, k)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to \code{2 \cdot \log(n)}, where \( n \) is the sample size.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the tail index estimator using the Weis method, based on differences in log-transformed order statistics.
}
\details{
This estimator calculates the tail index by analyzing the ratio of differences between selected order statistics:
- The default number of top-order statistics \( k \) is calculated as \( 2 \cdot \log(n) \).
- The tail index is computed as:
  \[
  \text{Tail Index} = \frac{\log(2)}{\log\left(\frac{X_{(k)} - X_{(1)}}{X_{(k/2)} - X_{(1)}}\right)}.
  \]
This method assumes that \( k \) is large enough to capture the tail behavior but small enough to avoid noise from the bulk of the distribution.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
Weis(x)

}
