% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_functions.R
\name{Zipf}
\alias{Zipf}
\title{Zipf Estimator (Quantile Plots Estimator)}
\usage{
Zipf(X, k, w)
}
\arguments{
\item{X}{A numeric vector containing the data sample.}

\item{k}{An optional integer specifying the number of top-order statistics to use. Defaults to the square root of the sample size.}

\item{w}{An optional vector of weights for the linear regression. Defaults to equal weights for all points.}
}
\value{
A numeric value representing the estimated tail index.
}
\description{
Computes the Zipf estimator, based on the Kratz and Resnick (1996) quantile plots approach, for the tail index of a heavy-tailed distribution.
}
\details{
The Zipf estimator uses a log-log plot of order statistics to estimate the tail index. A weighted linear regression is performed on the log-transformed data, where \code{w} specifies the weights for each point. The slope of the regression line is used to calculate the tail index.
}
\examples{
set.seed(123)
n = 1000
x = (abs(stats::rcauchy(n)))^(2)
Zipf(x)

}
